Update HtmlHelperExtensions.cs to include useful URL generation methods
As the URL's that come from the CMS are not routes, I find it easier to define two methods to help generate URL's to CMS pages so that I don't have to hard code them. The two methods posted help resolve this issue. Do we need to set up any examples of how to use these methods so that others are aware that they exist?